+++ examples/inputs-simple-sqlite/src/schema/__generated__/User/index.ts
@@ -1,3 +1,1 @@
 export * from './object.base';
-export * from './mutations';
-export * from './queries';