+++ src/crudGenerator/index.ts
@@ -12,18 +12,22 @@ export async function generateCrud(config: ConfigInternal, dmmf: DMMF.Document):
   const modelNames = dmmf.datamodel.models.map((model) => model.name);
 
   // Generate CRUD directories (e.g. User, Comment, ...)
-  await Promise.all(modelNames.map((model) => generateModel(config, dmmf, model)));
+  const models = await Promise.all(
+    modelNames.map(async (model) => {
+      const generated = await generateModel(config, dmmf, model);
+      return { model, generated };
+    }),
+  );
+
+  console.log(JSON.stringify({ models }));
 
   // Generate root objects.ts file (export all models + prisma objects)
-  const exports = dmmf.datamodel.models
-    .map((model) => `export * from './${model.name}';`)
-    .join('\n');
   const modelNamesEachLine = modelNames.map((model) => `'${model}',`).join('\n  ');
 
   await writeFile(
     config,
     'crud.objects',
-    useTemplate(objectsTemplate, { exports, ...config.crud, modelNames: modelNamesEachLine }),
+    useTemplate(objectsTemplate, { ...config.crud, modelNames: modelNamesEachLine }),
     path.join(config.crud.outputDir, 'objects.ts'),
   );
 
@@ -36,11 +40,37 @@ export async function generateCrud(config: ConfigInternal, dmmf: DMMF.Document):
   );
 
   // Generate root autocrud.ts file
+  // TODO REFACTOR AND TESTS
   if (config.crud.generateAutocrud) {
+    const imports = dmmf.datamodel.models
+      .map((model) => `import * as ${model.name} from './${model.name}';`)
+      .join('\n');
+
+    const modelsGenerated = dmmf.datamodel.models
+      .map((model) => {
+        const { name } = model;
+        return `  ${name}: {
+    Object: ${name}.${name}Object,
+    queries: ${(() => {
+      const queries =
+        models.find((el) => el.model === name)?.generated.filter((el) => el.type === 'queries') ||
+        [];
+      return `{\n${queries.map((el) => `      ${el.resolverName}: ${el.modelName}.${el.resolverName}${el.modelName}Query,`).join("\n")}\n    }`
+    })()},
+    mutations: ${(() => {
+      const mutations =
+        models.find((el) => el.model === name)?.generated.filter((el) => el.type === 'mutations') ||
+        [];
+      return `{\n${mutations.map((el) => `      ${el.resolverName}: ${el.modelName}.${el.resolverName}${el.modelName}Mutation,`).join("\n")}\n    }`
+    })()},
+  },`;
+      })
+      .join('\n');
+
     await writeFile(
       config,
       'crud.autocrud',
-      useTemplate(autoCrudTemplate, config.crud),
+      useTemplate(autoCrudTemplate, { ...config.crud, imports, modelsGenerated }),
       path.join(config.crud.outputDir, 'autocrud.ts'),
     );
   }