+++ src/crudGenerator/utils/generator.ts
@@ -2,15 +2,18 @@ import { DMMF } from '@prisma/generator-helper';
 import { ConfigInternal } from '../../utils/config';
 import { mutations as MutationTemplates } from '../templates/mutation';
 import { queries as QueryTemplates } from '../templates/query';
-import { writeIndex, writeObject, writeResolvers } from './parts';
+import { GeneratedResolver, writeIndex, writeObject, writeResolvers } from './parts';
 
+/**
+ * @returns List of generated resolvers
+ */
 export async function generateModel(
   config: ConfigInternal,
   dmmf: DMMF.Document,
   modelName: string,
-): Promise<void> {
+): Promise<GeneratedResolver[]> {
   const model = dmmf.datamodel.models.find((m) => m.name === modelName);
-  if (!model) return;
+  if (!model) return [];
 
   await writeObject(config, model);
   const queries = await writeResolvers(config, model, 'queries', QueryTemplates);
@@ -19,4 +22,6 @@ export async function generateModel(
     writeQueries: Boolean(queries.length),
     writeMutations: Boolean(mutations.length),
   });
+
+  return [...queries, ...mutations];
 }